requirements.txt
flask
PyJWT
pymongo